Yet another round of refactoring and moving functions around other modules.
Create a new module Peer that represents the interactions with the peer.
